requirements.txt
Django==5.2.1
djangorestframework==3.16.0
djangorestframework-api-response==0.1.0
django-redis==5.4.0
django-environ==0.9.0
django-ckeditor==6.3.2

pyotp==2.9.0
qrcode==8.2
pillow==12.0.0

uvicorn==0.34.2

channels==4.1.0
channels-redis==4.2.0

whitenoise==6.9.0

psycopg2==2.9.4
celery==5.5.3
django-celery-results==2.6.0
django-celery-beat==2.8.1

djoser==2.3.3
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
djangorestframework-simplejwt==5.5.1


django-storages==1.14.6
boto3==1.39.9
botocore==1.39.9
cryptography==45.0.5
rsa==4.9

django-axes==8.0.0
Faker==37.5.3
